+++ panel/widgets/tqdm_indicator.py
@@ -137,45 +137,3 @@ class Tqdm(pn.viewable.Viewer):
                 return _tqdm
 
         return ptqdm
-
-
-if __name__.startswith("bokeh"):
-    import time
-    import pandas as pd
-    import numpy as np
-
-    tqdm = Tqdm(layout="row", sizing_mode="stretch_width")
-
-    def run(*events):
-        for index in tqdm(range(0, 10)):
-            time.sleep(0.2)
-
-    button = pn.widgets.Button(name="Run Loop", button_type="primary")
-    button.on_click(run)
-
-    # Register Pandas. This gives DataFrame.progress_apply method
-    tqdm.tqdm.pandas(desc="my bar!")
-
-    df = pd.DataFrame(np.random.randint(0, 100, (100000, 6)))
-
-    def run_df(*events):
-        df.progress_apply(lambda x: x**2)
-
-    pandas_button = pn.widgets.Button(name="Pandas Apply", button_type="success")
-    pandas_button.on_click(run_df)
-    pandas_button
-
-    component = pn.Column(
-        button,
-        pandas_button,
-        tqdm,
-        sizing_mode="stretch_width"
-    )
-    template = pn.template.FastListTemplate(
-        title="Panel - Tqdm Indicator",
-        main=[component],
-        sidebar=[
-            pn.Param(tqdm, sizing_mode="stretch_width"),
-        ],
-    )
-    template.servable()
